requirements.txt
streamlit
langchain_groq
langchain_community
langchain
python-dotenv
arxiv
wikipedia
duckduckgo-search